+++ JGL/includes/jgl_json_handler.h
@@ -3,21 +3,21 @@
 
 #include "jgl_data.h"
 
-typedef 			void(*JSON_Funct)(std::fstream&, size_t, jgl::Data);
+typedef 			void(*JSON_Funct)(std::fstream&, uint32_t, jgl::Data);
 
 namespace jgl
 {
-	void json_add_line(std::fstream& file, const size_t tab, const jgl::String type_name, const jgl::String symbol);
-	void json_add_line(std::fstream& file, const size_t tab, const jgl::String text);
+	void json_add_line(std::fstream& file, const uint32_t tab, const jgl::String type_name, const jgl::String symbol);
+	void json_add_line(std::fstream& file, const uint32_t tab, const jgl::String text);
 
 
 	template <typename T>
-	void json_add_vector(std::fstream& file, const size_t tab, const jgl::String type_name, const jgl::Array<T>& vector, const JSON_Funct p_funct, Data p_data)
+	void json_add_vector(std::fstream& file, const uint32_t tab, const jgl::String type_name, const jgl::Array<T>& vector, const JSON_Funct p_funct, Data p_data)
 	{
 		Data data;
 
 		json_add_line(file, tab, type_name, "[");
-		for (size_t i = 0; i < vector.size(); i++)
+		for (uint32_t i = 0; i < vector.size(); i++)
 		{
 			data = p_data + Data(&(vector[i]));
 			json_add_line(file, tab + 1, "{");
@@ -34,7 +34,7 @@ namespace jgl
 	}
 
 	template <typename T, typename V>
-	void json_add_map(std::fstream& file, const size_t tab, const jgl::String type_name, const std::map<T, V> map, const JSON_Funct p_funct, Data p_data)
+	void json_add_map(std::fstream& file, const uint32_t tab, const jgl::String type_name, const std::map<T, V> map, const JSON_Funct p_funct, Data p_data)
 	{
 		if (map.size() == 0)
 			return;
