+++ JGL/includes/structure/jgl_plot.h
@@ -8,7 +8,7 @@ namespace jgl
 		jgl::String(*funct)(float);
 		float min;
 		float max;
-		size_t precision;
+		uint32_t precision;
 		float delta;
 
 		Plot_data(jgl::String p_name = "Value", float p_min = 0.0f, float p_max = 100.0f, jgl::String(*p_funct)(float) = nullptr)
@@ -27,7 +27,7 @@ namespace jgl
 		void set_min(float p_min){min = p_min;delta = (max - min) / 10;}
 		void set_max(float p_max){max = p_max;delta = (max - min) / 10;}
 		void set_funct(jgl::String(*p_funct)(float)){funct = p_funct;}
-		void divide(size_t nb_division){delta = (max - min) / nb_division;}
+		void divide(uint32_t nb_division){delta = (max - min) / nb_division;}
 		void set_gap(float p_delta){delta = p_delta;}
 		float range(){return (max - min);}
 	};
@@ -64,8 +64,8 @@ namespace jgl
 		Vector2 _pos_up;
 		Vector2 _pos_right;
 
-		size_t _point_size;
-		size_t _line_size;
+		uint32_t _point_size;
+		uint32_t _line_size;
 
 		std::vector<Line> _lines;
 
@@ -98,12 +98,12 @@ namespace jgl
 		void calc_axis_unit();
 
 		void add_line(Color p_color = Color(0, 0, 0));
-		void add_point(Vector2 p_point, size_t line = 0);
-		void add_point(float p_a, float p_b, size_t line = 0){add_point(Vector2(p_a, p_b), line);}
-		void set_line_color(size_t index, Color p_color);
+		void add_point(Vector2 p_point, uint32_t line = 0);
+		void add_point(float p_a, float p_b, uint32_t line = 0){add_point(Vector2(p_a, p_b), line);}
+		void set_line_color(uint32_t index, Color p_color);
 
-		void set_point_size(size_t p_size){_point_size = p_size;}
-		void set_line_size(size_t p_size){_line_size = p_size;}
+		void set_point_size(uint32_t p_size){_point_size = p_size;}
+		void set_line_size(uint32_t p_size){_line_size = p_size;}
 
 		void set_size(Vector2 p_size){_size = p_size;actualize_point();}
 		void set_margin(Vector2 p_margin){_margin = p_margin;actualize_point();}
@@ -117,7 +117,7 @@ namespace jgl
 		void set_absciss_min(float p_min){absciss().set_min(p_min);calc_axis_unit();}
 		void set_absciss_max(float p_max){absciss().set_max(p_max);calc_axis_unit();}
 		void set_absciss_gap(float p_gap){absciss().set_gap(p_gap);calc_axis_unit();}
-		void divide_absciss(size_t subdivision){absciss().divide(subdivision);calc_axis_unit();}
+		void divide_absciss(uint32_t subdivision){absciss().divide(subdivision);calc_axis_unit();}
 		void set_absciss_precision(int p_precision){absciss().precision = p_precision;}
 
 		void set_ordinate_name(jgl::String p_name){ordinate().name = jgl::String(p_name);}
@@ -125,7 +125,7 @@ namespace jgl
 		void set_ordinate_min(float p_min){ordinate().set_min(p_min);calc_axis_unit();}
 		void set_ordinate_max(float p_max){ordinate().set_max(p_max);calc_axis_unit();}
 		void set_ordinate_gap(float p_gap){ordinate().set_gap(p_gap);calc_axis_unit();}
-		void divide_ordinate(size_t subdivision){ordinate().divide(subdivision);calc_axis_unit();}
+		void divide_ordinate(uint32_t subdivision){ordinate().divide(subdivision);calc_axis_unit();}
 		void set_ordinate_precision(int p_precision){ordinate().precision = p_precision;}
 	};
 }
